商品情報削除機能修正1回目

+++ app/controllers/items_controller.rb
@@ -1,5 +1,5 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: [:new, :create, :edit, :update]
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
   before_action :move_to_index, only: [:edit, :update]
   before_action :item_params, only: [:edit, :show, :update]
 
@@ -36,8 +36,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user_id == current_user.id
+      item.destroy
+      redirect_to root_path
+    else
+      render :show
+    end
   end
 
   private
